Add files via upload
Introduction
The newly developed sensors show outliers in the data. The reason why the outliers are reported is not fully clear, but it is important to have tools that can tell us if any changes applied actually improve the output.
Task: Outlier Detection
Sample data will be made available in the form of a pickled pandas DataFrame. The task is to output an outlier frequency per sensor type and sensor unit, e.g.:
SensorA:
CO2: x outliers/h
PM1.0: y outliers/h
The script should be using pandas and json as dependencies (imports), but preferably no other dependencies. For development it is naturally fine to use matplotlib for plotting,